Gate and Policies added

+++ app/Http/Controllers/CommentController.php
@@ -3,17 +3,25 @@
 namespace App\Http\Controllers;
 
 use App\Models\Comment;
+// use Illuminate\Auth\Access\Gate;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 
 class CommentController extends Controller
 {
     //
 
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function create(){
 
         $comment = new Comment();
         $comment->content = request()->content;
         $comment->article_id = request()->article_id;
+        $comment->user_id = auth()->user()->id;
         $comment->save();
 
         return back();
@@ -22,8 +30,15 @@ class CommentController extends Controller
     public function delete($id)
     {
         $comment = Comment::find($id);
-        $comment->delete();
 
-        return back()->with('info','A comment deleted');
+        if(Gate::allows('comment-delete', $comment)){
+            $comment->delete();
+            return back()->with('info','A comment deleted');
+
+        }else{
+            return back()->with('info','Unauthorize');
+        }
+        
+
     }
 }
